+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/FrameworkExtensionTest.php
@@ -253,6 +253,24 @@ abstract class FrameworkExtensionTest extends TestCase
         $this->assertFalse($container->hasDefinition('data_collector.config'), '->registerProfilerConfiguration() does not load collectors.xml');
     }
 
+    public function testProfilerCollectSerializerDataEnabled()
+    {
+        $container = $this->createContainerFromFile('profiler_collect_serializer_data');
+
+        $this->assertTrue($container->hasDefinition('profiler'));
+        $this->assertTrue($container->hasDefinition('serializer.data_collector'));
+        $this->assertTrue($container->hasDefinition('debug.serializer'));
+    }
+
+    public function testProfilerCollectSerializerDataDefaultDisabled()
+    {
+        $container = $this->createContainerFromFile('profiler');
+
+        $this->assertTrue($container->hasDefinition('profiler'));
+        $this->assertFalse($container->hasDefinition('serializer.data_collector'));
+        $this->assertFalse($container->hasDefinition('debug.serializer'));
+    }
+
     public function testWorkflows()
     {
         $container = $this->createContainerFromFile('workflows');